backend/pyproject.toml
[tool.poetry]
name = "nse-trader-backend"
version = "0.1.0"
description = "FastAPI backend for NSE Trader platform"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app", from = "."}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = {extras = ["email"], version = "^2.0.0"}
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
redis = {extras = ["hiredis"], version = "^5.0.0"}
celery = {extras = ["redis"], version = "^5.3.0"}
pika = "^1.3.0"
requests = "^2.31.0"
pandas = "^2.1.0"
numpy = "^1.26.0"
tradingview-ta = "^3.3.0"  # TradingView technical analysis

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.25.0" # For async testing FastAPI endpoints

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
